backend/scrape-and-process-listings/email-alerts/requirements.txt
firebase-admin==6.9.0
google-cloud-firestore==2.21.0
python-dotenv==1.1.1
